[1183] Ticket #68: amd64fix

+++ src/game/UpdateData.cpp
@@ -117,7 +117,7 @@ bool UpdateData::BuildPacket(WorldPacket *packet)
     if (m_data.size() > 50 )
     {
 
-        unsigned long destsize = buf.size() + buf.size()/10 + 16;
+        uint32 destsize = buf.size() + buf.size()/10 + 16;
         packet->resize( destsize );
 
         packet->put(0, (uint32)buf.size());
